fix(scheduled-tasks): address PR review comments
- Move ListScheduledTasksParams/ListScheduledTaskRunsParams out of
  scheduled-tasks.api.ts into models/scheduled-tasks.ts, matching the
  repo convention of keeping shared interfaces in a models directory.
- Replace the hand-rolled box-shadow on ScheduledTaskDetailView's
  History card with the existing Tailwind shadow-md utility, which
  already matches the same two-layer shadow values.
