+++ apps/chat/src/models/scheduled-tasks.ts
@@ -1,0 +1,18 @@
+import type { ListScheduledTasksSortEnum } from '@epam/chat-api-client';
+
+/** Params for `listScheduledTasks`. */
+export interface ListScheduledTasksParams {
+  limit?: number;
+  offset?: number;
+  search?: string;
+  sort?: ListScheduledTasksSortEnum;
+  signal?: AbortSignal;
+}
+
+/** Params for `listScheduledTaskRuns`. */
+export interface ListScheduledTaskRunsParams {
+  scheduleId: string;
+  limit?: number;
+  offset?: number;
+  signal?: AbortSignal;
+}
