+++ libs/scheduled-tasks/src/components/ScheduledTaskDetailView/ScheduledTaskDetailView.module.scss
@@ -34,7 +34,4 @@
 
 .historyCard {
   background: var(--stdv-history-bg, var(--bg-layer-raised, #fcfcfc));
-  box-shadow:
-    0 6px 16px 0 rgba(39, 100, 217, 0.06),
-    0 6px 24px 0 rgba(22, 27, 45, 0.03);
 }
